// src/utils/web-rtc/interface.ts
export interface WebRTC {
  MediaStream: typeof MediaStream;
  MediaStreamTrack: typeof MediaStreamTrack;
  RTCDataChannel: typeof RTCDataChannel;
  RTCDataChannelEvent: typeof RTCDataChannelEvent;
  RTCDtlsTransport: typeof RTCDtlsTransport;
  RTCIceCandidate: typeof RTCIceCandidate;
  RTCIceTransport: typeof RTCIceTransport;
  RTCPeerConnection: typeof RTCPeerConnection;
  RTCPeerConnectionIceEvent: typeof RTCPeerConnectionIceEvent;
  RTCRtpReceiver: typeof RTCRtpReceiver;
  RTCRtpSender: typeof RTCRtpSender;
  RTCRtpTransceiver: typeof RTCRtpTransceiver;
  RTCSessionDescription: typeof RTCSessionDescription;
}